tests/firmware/requirements.txt
# Python dependencies for RetroCPU firmware tests

# Core testing framework
pytest>=7.4.0
pytest-timeout>=2.1.0

# Serial communication
pyserial>=3.5

# Optional: test reporting
pytest-html>=3.2.0
pytest-json-report>=1.5.0
